Update patch set 3
Patch Set 3: Code-Review+1

Perfectly done. This is the most clean way as this doesn't leave any unused cruft in the gproj.mk file. To me it looks okay and hopefully somebody will merge this soon as this fixes the build process for cm-12.1 again for the e975. I add some people to the list who can merge this.

Patch-set: 3
Label: Code-Review=+1